Moved gp module to be part of gp_forecaster.
