+++ test/model/gp_forecaster/test_inference.py
@@ -16,7 +16,7 @@ import numpy as np
 
 # First-party imports
 from gluonts.kernels import RBFKernel
-from gluonts.gp import GaussianProcess
+from gluonts.model.gp_forecaster.gaussian_process import GaussianProcess
 
 # Relative imports
 from .data import (
